LPS-172977 Move UnsupportedBusinessTypes to api module and expose it in the bnd
