+++ modules/apps/object/object-rest-api/src/main/resources/com/liferay/object/rest/odata/entity/v1_0/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
